SP5 WPF: WpfPlot → WpfPlotGL
