+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.WPF/MainWindow.xaml
@@ -8,5 +8,5 @@
         mc:Ignorable="d"
         WindowStartupLocation="CenterScreen"
         Title="ScottPlot 5 - WPF Sandbox" Height="300" Width="500">
-    <ScottPlot:WpfPlot x:Name="WpfPlot1" />
+    <ScottPlot:WpfPlotGL x:Name="WpfPlot1" />
 </Window>
